Handle share link for authorized user  (#358)
* refactor(Login): move to `navigationGuard`

ADAMANTPWA-443
- Move navigate by URI logic to separate reusable method.
- Move `navigateByContact` to `navigationGuard`.

* feat(isLogged.js): redirect to Chats

ADAMANTPWA-443
- Redirect to Chats if current URI not containing contact information.
