+++ src/middlewares/isLogged.js
@@ -1,8 +1,9 @@
+import { navigateByURI } from '@/router/navigationGuard'
 import store from '@/store'
 
 export default (to, from, next) => {
   if (to.name === 'Login' && store.getters.isLogged) {
-    next({ name: 'Chats' })
+    navigateByURI(next)
   } else {
     next()
   }
